requirements.txt
transformers==4.35.2
pandas
numpy
scikit-learn
matplotlib
seaborn
streamlit
tmdbv3api
python-dotenv